tech(WebSocket) Adding SocketIO
This cl adds SocketIO and wires up a simple "chat room" infrastructure.
The primary motivation for this is to pave some of the plumbing to set
up a way for RTCPeerConnections to get established between one another.

This will be broken down in a multi-part process. The first part was
this cl, adding websockets. After the websockets I'll add the userMedia
portion (which won't be hard.), followed by establishing a connection
between RTCPeerConnections over websockets (I haven't done this before,
so I might go off on a tangent before I find the "right" way), and
finally setup a signaling layer.

With any luck I might be able to toss this up in a server to use for
reals, but thats a good chunk in the future.