tech(RTCPeerConnection) PeerConnection barebones
This cl introduces barebone peer connections. With this, I should have
the absolute minimum required to establish sending/receiving UserMedia
with WebRTC.

The general flow is:

Local

```
createPeerConnection
getUserMedia()
  createOffer()
    setLocalDescription
    emit('offer') // with websocket
```

Remote

```
receiveOffer() // with websocket
  setRemoteDescription()
  getUserMedia()
    createAnswer()
      setLocalDescription()
      emit('answer') // with websocket
```

Local

```
receiveAnswer() // with websocket
  setRemoteDescription()
```

Along with this flow, `iceCandidate` information is passed around
through websocket messages as well.

This approach is riddled with bugs, but I wanted focus on making a
smaller cl than something more fleshed out.

Following this I will either be tossing this on aws, or setup STUN
and/or TURN servers.