docs(*) Updating docs
This cl updates the documentation of this application. The previous cl
mentioned there was a 502 error. The error was resolved by making
configuration changes on AWS.

After successfully deploying the app I ran into a slew of bugs. Nearly
all the bugs I found were resolved by changing configuration with AWS
or my host provider, very little "code" changes were necessary. But, I
still wanted to document what went wrong so that maybe I can save myself
a little time in the future.

**Whats left**:

I'm able to establish a connection with myself only... which is
frightenly not useful. The next logic step is to setup a TURN and/pr
STUN server so that I can successfully make video calls with people not
myself.

Stay tuned! :)