fix/design(*) Fix all the things!!
This cl does a large set of updates, both fixing broken behavior and
improving the overall UI of the site.

The root cause of the failure of the peer connections not sending
userMedia was a combination of not having the shim library `adapter.js`
and not including `STUN` servers. After including the shim and adding a
few stun severs from google... everything simply worked.

In the process of debugging, I added a datachannel layer, so that I
could test if the RTCPeerConnections were actually established for real.
I didn't need to keep the DataChannel text fields... but it was
convenient to debug the connection, so I kept it for continued
debugging.

As for design, I separated the site into discrete sections:

* videos
* websocket messaging
* datachannel messaging

Having a clean division between everything helped make the page feel
less cluttered/chaotic.

With this... I have a working video chat application. It doesn't work
100% (of course)... two problems I noticed was:

* iphone was sending out video, but was not displaying video
* android was rendering two video streams for some reason.

So yeah, I want to go back and fix those issues when I get the chance.
Also, it would be better if I was using my own stun server (mostly for
learning purposes... but also because its not necessarily safe to use a
different hosts' stun servers.

After reading into this a bit more, I'm really, really not feeling like
making a Turn server. it sounds like a more costly option that I don't
necessarily want to go down. Maybe in the future, but theres other
things I'd like to explore instead.