fix(MediaStream) iOS video
This cl fixes an issue where video was not working on for iPhones. The
root cause was iPhone needed a `playsinline` attribute in order for the
video to play... inline.

Also, I fixed a cosmetic issue where the local video was "wrong". While
testing this, I was given feedback that the local video's normal display
should be given matching movement when moving, as if I was looking at a
mirror. This fix was as simple as adding a line of css to `transform`.

https://webkit.org/blog/6784/new-video-policies-for-ios/