tech(Architecture) VueJs
This cl is part of an ongoing process to componentize the different
layers of this application. Ultimately I want to split up the
application into the following components:

* view - rendering component, consumers the other modules
* signaling - module dedicated to connection hand shakes
* connection - module dedicated to managing peer connections and ice
  candidates... as well as handling userMedia

The view will be able to invoke methods from the signaling layer, but
will also listen to events emited from signaling (such as a user
joining, or if an offer is requested.)

Also I'll be setting up a conference type of view, but thats for the
future.