tech(Webpack) Setting up webpack
This cl adds a barebones webpack setup to the application. Following
this cl, I'll be breaking apart `public/script.js` into smaller
components that can get reused for different pages.

I mentioned earlier that the plan is to have `view`, `signaling` and
`connection` components... the current plan is to swap out the view
based on what page the user is in... either a page dedicated to video
chatting with currently connected users, or video conference.