tech(Architecture) Split into smaller components
This cl splits the previous `script.js` into multiple smaller
components. While this wasn't entirely necessary, it does help in
drawing out responsibilities of the different layers of communication.

With this exercise we can see that the `connection` layer's entry points
are `createOffer` and `acceptOffer`.

The `signaling` layer is a little more involved, instead the webrtc entry
points appear to be `sendOffer`, `sendAnswer`, and `sendIceCandidate`
(... also `sendMessage`, but thats for debugging).

So... yeah! This was a fun exercise. If I want to make a conference
room, I'd just need to add a different viewModel that auto-accepts
peerConnection offers, so maybe thats next.

Added bonus: this cl also fixes an issue where things would
intermittently break if I opened and closed peerConnections enough. I've
changed the when/how peerConnections get created.  Previously there was
always a peerConnection that would get created before sending or
receiving an offer. Once an offer was accepted, then a new
peerConnection was created in place of the old. This was the root cause
of some strange behavior where media would sometimes not get included...

The fix was to simply change when peerConnections are created. With this
cl, peerConnections are only created when sending or receiving an offer,
and are stored in a map when an answer is dispatched. By doing this,
the chances of incorrectly answering, or not adding IceCandidates or
mediaTracks gets dropped altogether. Poking around I am able to have
multiple peerConnections open at once, have some drop, and others get
added, all without (noticable) bugs, which is exciting because setting
up a conference scenario will be even easier.